+++ ntex/src/http/builder.rs
@@ -90,7 +90,7 @@ where
     /// To disable timeout set value to 0.
     ///
     /// By default disconnect timeout is set to 3 seconds.
-    pub fn client_disconnect(mut self, val: u64) -> Self {
+    pub fn disconnect_timeout(mut self, val: u64) -> Self {
         self.client_disconnect = val;
         self
     }